requirements.txt
pandas==2.3.0
numpy==2.3.0
matplotlib==3.10.3
seaborn==0.13.2
ipykernel==6.29.5
pathlib==1.0.1
importlib==1.0.4
pillow==11.2.1
moviepy==2.2.1